// demo-mcp-21-code-generator-node/input/reference-repos/ECOMMERCE_V9/MSAbCaAlarPricesETL/src/common/utils/enums/taks.enum.ts
export enum Etask {
  CREATE = 'PROCESS_CREATE',
  BEGIN_JOB = 'BEGIN_JOB',
  FINDALL = 'CONSULTANDO_MOCKUPS',
  FINDONE = 'CONSULTANDO_MOCKUP',
  UPDATE = 'ACTUALIZANDO_MOCKUP',
  REMOVE = 'ELIMINANDO_MOCKUP',
  LOAD_MESSAGE = 'CARGANDO_MENSAJES',
  LOAD_PARAM = 'CARGANDO_PARAMETROS',
  ERROR_LOAD_PARAM = 'ERROR_CARGANDO_PARAMETROS',
  CHANNEL = 'VALIDATE_CHANNEL',
  CONSUMED_SERVICE = 'CONSUMED_SERVICE',
  PROCESS_KAFKA_ACCEPT_RETRY_TOPIC = 'PROCESS_KAFKA_ACCEPT_RETRY_TOPIC',
  PROCESS_KAFKA_START_ORDER_RETRY_TOPIC = 'PROCESS_KAFKA_START_ORDER_RETRY_TOPIC',
  PROCESS_KAFKA_RESERVE_NUMBER_RETRY_TOPIC = 'PROCESS_KAFKA_RESERVE_NUMBER_RETRY_TOPIC',
  PROCESS_KAFKA_PAYMENT_CONF_RETRY_TOPIC = 'PROCESS_KAFKA_PAYMENT_CONF_RETRY_TOPIC',
  PROCESS_KAFKA_UPDATE_BILL_MED_RETRY_TOPIC = 'PROCESS_KAFKA_UPDATE_BILL_MED_RETRY_TOPIC',
  PROCESS_KAFKA_RESERVE_NUMBER_CONFIRM_RETRY_TOPIC = 'PROCESS_KAFKA_RESERVE_NUMBER_CONFIRM_RETRY_TOPIC',
  PROCESS_KAFKA_CONTRACT_RETRY_TOPIC = 'PROCESS_KAFKA_CONTRACT_RETRY_TOPIC',
  PROCESS_KAFKA_UPDATE_ITEM_RETRY_TOPIC = 'PROCESS_KAFKA_UPDATE_ITEM_RETRY_TOPIC',
  PROCESS_KAFKA_SUBMIT_RETRY_TOPIC = 'PROCESS_KAFKA_SUBMIT_RETRY_TOPIC',
  PROCESS_KAFKA_PAYMENT_RETRY_TOPIC = 'PROCESS_KAFKA_PAYMENT_RETRY_TOPIC',
  EXCEPTION_MANAGER = 'EXCEPTION_MANAGER',
  VALIDATE_REQUEST = 'VALIDATE_REQUEST',
  HTTP_RESPONSE = 'HTTP_RESPONSE',
  EMIT_KAFKA = 'EMITIENDO EVENTO KAFKA',
  FIND_GETPAGES = 'CONSULTANDO GETPAGES',
  FIND_FINDLASTPRICES = 'CONSULTANDO FINDLASTPRICES',
  FIND_GETDATACATALOG = 'CONSULTANDO GETDATACATALOG',
  GET_PRODUCT_OFFERING_DATA = 'GET_PRODUCT_OFFERING_DATA',
  ERROR_CREATE_TRACEABILITY = 'ERROR_CREATE_TRACEABILITY',
  ERROR_TO_CREATE_SERVICE_ERROR = 'ERROR_TO_CREATE_SERVICE_ERROR',
  DELETE_DATA = 'DELETE_DATA',
  SAVE_DATA = 'SAVE_DATA',
  APM='CAPTURE_APM_QUERY',
}

export enum ETaskPrices {
  ERROR_SERVICE = 'Error en el services impl',
  ERROR_DELETE_DATABASE_EQUIPMENT = 'Error al eliminar la base de datos Equipment',
  ERROR_CONSULT_COLLECTION_COLPRTPRODUCTOFFERING = 'Error al consultar la coleccion COLPRTPRODUCTOFFERING',
  ERROR_PRICE_EQUIPMENTUC = 'Error en el metodo terminales',
  ERROR_PRICE_EQUIPMENTPROCESSUC = 'Error en el metodo pricesEquipmentProcess',
  ERROR_PRICE_PROCESSLASTPRICESUC = 'Error en el metodo terminales Last Prices',
  ERROR_PRICE_TECHNOLOGYUC = 'Error en el metodo tecnologia',
  ERROR_PRICE_TECHNOLOGYPROCESSUC = 'Error en el metodo pricesTechnologyProcess',
  ERROR_PRICE_POSPAGOUC = 'Error en el metodo pospago',
  ERROR_PRICE_PREPAGOUC = 'Error en el metodo prepago',
  ERROR_PRICE_MOBILEPROCESSUC = 'Error en el metodo moviles',
  ERROR_PRICE_HOMEUC = 'Error en el metodo hogares',
  ERROR_PRICE_HOMEPROCESSUC = 'Error en el metodo pricesHomeProcess',
  ERROR_CONNECTION_DATABASE = 'Error en la conexion con la base de datos',
  ERROR_SAVE_DATABASE = 'Error en el guardado de datos de Equipment',
  ERROR_GET_DATA_CATALOG = 'Error al obtener la data del valgenerate',
  ERROR_GET_PAGES = 'Error al obtener el numero de paginas',
  ERROR_FIND_LAST_PRICES = 'Error al obtener la data de base de datos "LastPrices"',
  ERROR_PROCESS_KAFKA = 'Error al emitir el evento kafka',
  ERROR_CLOSE_DATABASE = 'Error en el cerrado de la base de datos',
  ERROR_TO_SAVE_DATA = 'Error para guardar la data',
  ERROR_TO_DELETE_DATA = 'Error para borrar la data',
  ERROR_CREATE_TRACEABILITY = 'Error al guardar la traceability'
}

//Descripción de las tareas que se realizan en el microservicio
export enum ETaskDesc {
  CHANNEL = 'Validation of the channel',
  ERROR_LOAD_PARAM = 'Error cargando parametros',
  UPDATE_PARAM = 'Actualizando parametros',
  ERROR_UPDATE_PARAM = 'Error actualizando parametros',
  ERROR_LOAD_MESSAGES = 'Error cargando mensajes',
  ERROR_UPDATE_MESSAGES = 'Error actualizando mensajes',
  CONSUMED_SERVICE = 'Result service',
  CONTADOR = 'Numero de paginas',
  FINISH_PAGE = 'Finalizo en la pagina',
  START_EQUIPMENT = 'Inicia terminales',
  FINSH_EQUIPMENT = 'Finalizo terminales',
  START_TECHNOLOGY = 'Inicia tecnologia',
  FINSH_TECHNOLOGY = 'Finalizo tecnologia',
  START_POSPAGO = 'Inicia pospago',
  FINSH_POSPAGO = 'Finalizo pospago',
  START_PREPAGO = 'Inicia prepago',
  FINSH_PREPAGO = 'Finalizo prepago',
  START_HOME = 'Inicia hogares',
  FINSH_HOME = 'Finalizo hogares',
  START_SERVICE = 'Inicia servicio',
  FINSH_SERVICE_PRICE = 'Finalizo servicio',
  START_GET_DATA_PRODUCT_OFFERING = 'Start get product offering data',
  FINISH_GET_DATA_PRODUCT_OFFERING = 'Finish get product offering data',
  ERROR_TO_CREATE_TRACEABILITY = 'Error al crear data de traceabilidad',
  ERROR_TO_CREATE_SERVICE_ERROR = 'Error al crear data para service error',
  START_DELETE_DATA = 'Start delete data',
  FINISH_DELETE_DATA = 'Finish delete data',
  START_SAVE_DATA = 'Start save data',
  FINISH_SAVE_DATA = 'Finish save data',
  START_SAVE_DATA_LAST_PRICES = 'Start save data to last prices',
  FINISH_SAVE_DATA_LAST_PRICES = 'Finish save data to last prices',
  START_FLOW_TO_SAVE_DATA = 'Start flow to save data'
}